eliminate custom build and test steps
